refactor(dynamic-agents): split agent_runtime.py into focused modules
Extract from the 1224-line agent_runtime.py into three new modules:
- skills.py (138 lines): skill loading + llm_prompt fallback
- streaming.py (377 lines): StreamingMixin with stream/resume/interrupt
- pool.py (176 lines): AgentRuntimeCache + get_runtime_cache singleton

agent_runtime.py drops to 590 lines focused on init, tools, and
subagent wiring. All external imports (get_runtime_cache, etc.) are
re-exported for backward compatibility.
